analyse_level parameter added - wrting bout model and layer stats

+++ edgeai_tidlrunner/runner/modules/vision/pipelines/compile_/analyze.py
@@ -44,6 +44,13 @@ from . import compile
 from . import infer
 
 
+def _write_outputs_to_bin(root, basename, output_dict):
+    os.makedirs(root, exist_ok=True)
+    for output_name, output_tensor in output_dict.items():
+        out = np.array(output_tensor, dtype=np.float32)
+        out.tofile(os.path.join(root, basename + output_name.replace("/", "_") + '.bin'))
+
+
 class CompileAnalyzeNoTIDL(compile.CompileModel):
     ARGS_DICT=SETTINGS_DEFAULT['analyze']
     COPY_ARGS=COPY_SETTINGS_DEFAULT['analyze']
@@ -57,18 +64,19 @@ class CompileAnalyzeNoTIDL(compile.CompileModel):
 
     def modify_model(self):
         super().modify_model()
-        model_path = self.model_path
-        # Load the original ONNX model and add intermediate outputs
-        onnx_model = onnx.load(model_path)
-        intermediate_layer_value_info = onnx.helper.ValueInfoProto()
-        intermediate_layer_value_info.name = ''
-        for i in range(len(onnx_model.graph.node)):
-            for j in range(len(onnx_model.graph.node[i].output)):
-                intermediate_layer_value_info.name = onnx_model.graph.node[i].output[j]
-                onnx_model.graph.output.append(intermediate_layer_value_info)
+        if self.kwargs['common.analyze_level'] > 0:
+            model_path = self.model_path
+            # Load the original ONNX model and add intermediate outputs
+            onnx_model = onnx.load(model_path)
+            intermediate_layer_value_info = onnx.helper.ValueInfoProto()
+            intermediate_layer_value_info.name = ''
+            for i in range(len(onnx_model.graph.node)):
+                for j in range(len(onnx_model.graph.node[i].output)):
+                    intermediate_layer_value_info.name = onnx_model.graph.node[i].output[j]
+                    onnx_model.graph.output.append(intermediate_layer_value_info)
+                #
             #
-        #
-        onnx.save(onnx_model, model_path)
+            onnx.save(onnx_model, model_path)
 
     def _prepare(self):     
         if os.path.exists(self.run_dir):
@@ -95,19 +103,13 @@ class InferAnalyzeNoTIDL(infer.InferModel):
 
     def _run(self):
         super()._run()
-        traces_root = os.path.join(self.run_dir, 'traces')
+        traces_root = os.path.join(self.run_dir, 'outputs_')
         os.makedirs(traces_root, exist_ok=True)
         for frame_idx in range(len(self.run_data)):
             frame_name = str(frame_idx)
-            self.write_outputs_to_bin(os.path.join(traces_root, frame_name), '', self.run_data[frame_idx]['output'])
+            _write_outputs_to_bin(os.path.join(traces_root, frame_name), '', self.run_data[frame_idx]['output'])
         #
-        print('INFO: onnxruntime traces generated')
-
-    def write_outputs_to_bin(self, root, basename, output_dict):
-        os.makedirs(root, exist_ok=True)
-        for output_name, output_tensor in output_dict.items():
-            out = np.array(output_tensor, dtype=np.float32)
-            out.tofile(os.path.join(root, basename + output_name.replace("/", "_") + '.bin'))
+        print('INFO: onnxruntime outputs generated')
 
 
 class CompileAnalyzeTIDL(compile.CompileModel):
@@ -131,20 +133,31 @@ class InferAnalyzeTIDL(infer.InferModel):
     def __init__(self, **kwargs):
         kargs_copy = copy.deepcopy(kwargs)
         kargs_copy['session.run_dir'] = os.path.join(kargs_copy['session.run_dir'], 'tidl')
-        kargs_copy['session.runtime_options.debug_level'] = 4
+        kargs_copy['session.runtime_options.debug_level'] = 0 if kargs_copy['common.analyze_level'] == 0 else 4
         kargs_copy['common.postprocess_enable'] = False            
         super().__init__(**kargs_copy)
 
     def _run(self):
         super()._run()
+        traces_root = os.path.join(self.run_dir, 'outputs_')
+        os.makedirs(traces_root, exist_ok=True)
+        for frame_idx in range(len(self.run_data)):
+            frame_name = str(frame_idx)
+            _write_outputs_to_bin(os.path.join(traces_root, frame_name), '', self.run_data[frame_idx]['output'])
+        #
+        print('INFO: TIDL outputs generated')
         print('INFO: TIDL traces generated')
 
     def _run_frame(self, input_index, *args, **kwargs):
-        os.system('rm -f /tmp/tidl_*.bin')
-        run_dict = super()._run_frame(input_index, *args, **kwargs)
-        traces_root = os.path.join(self.run_dir, 'traces', str(input_index))
-        os.makedirs(traces_root, exist_ok=True)
-        os.system(f'mv /tmp/tidl_*.bin {traces_root}')
+        if self.kwargs['common.analyze_level'] == 0:
+            run_dict = super()._run_frame(input_index, *args, **kwargs)
+        else:
+            os.system('rm -f /tmp/tidl_*.bin')
+            run_dict = super()._run_frame(input_index, *args, **kwargs)
+            traces_root = os.path.join(self.run_dir, 'traces_', str(input_index))
+            os.makedirs(traces_root, exist_ok=True)
+            os.system(f'mv /tmp/tidl_*.bin {traces_root}')
+        #
         return run_dict
 
 
@@ -158,37 +171,96 @@ class InferAnalyzeFinal(compile_base.CompileModelBase):
     def _run(self):
         notidl_run_dir = os.path.join(self.run_dir, 'notidl')
         tidl_run_dir = os.path.join(self.run_dir, 'tidl')
-        notidl_traces = os.path.join(notidl_run_dir, 'traces')
-        tidl_traces = os.path.join(tidl_run_dir, 'traces')
-        layer_info_dir = os.path.join(tidl_run_dir, 'artifacts', 'tempDir')
-        layer_info_files = [f for f in os.listdir(layer_info_dir) if f.endswith('layer_info.txt')]
-        layer_info_path = os.path.join(layer_info_dir, layer_info_files[0])
+        notidl_outputs_dir = os.path.join(notidl_run_dir, 'outputs_')
+        tidl_traces_dir = os.path.join(tidl_run_dir, 'traces_')
+        tidl_outputs_dir = os.path.join(tidl_run_dir, 'outputs_')
 
+        notidl_outputs_dirs = [os.path.join(notidl_outputs_dir,f) for f in os.listdir(notidl_outputs_dir)]
+        tidl_traces_dirs = [os.path.join(tidl_traces_dir,f) for f in os.listdir(tidl_traces_dir)]
+        tidl_outputs_dirs = [os.path.join(tidl_outputs_dir,f) for f in os.listdir(tidl_outputs_dir)]
+
+        notidl_outputs_dirs.sort()
+        tidl_traces_dirs.sort()
+        tidl_outputs_dirs.sort()
+
+        if len(notidl_outputs_dirs) != len(tidl_outputs_dirs):
+            print(f"ERROR: Number of outputs in notidl and tidl folders do not match")
+            raise ValueError(f"ERROR: Number of outputs in notidl and tidl folders do not match")
+        #
+        
         analyze_xlsx_path = os.path.join(self.run_dir, "analyze.xlsx")
         analyze_xlsx =  xlsxwriter.Workbook(analyze_xlsx_path, options=dict(nan_inf_to_errors=True))
-        num_traces = len(os.listdir(notidl_traces))
-        for frame_idx in range(num_traces):
-            frame_worksheet = analyze_xlsx.add_worksheet(str(frame_idx))
-            notidl_traces_frame_id_dir = os.path.join(notidl_traces, str(frame_idx))
-            tidl_traces_frame_id_dir = os.path.join(tidl_traces, str(frame_idx))
-            tidl_onnx_trace_mapping = self.get_traces(notidl_traces_frame_id_dir, tidl_traces_frame_id_dir, layer_info_path)
-            layers_diff = self.get_layers_diff(notidl_traces_frame_id_dir, tidl_traces_frame_id_dir, layer_info_path, tidl_onnx_trace_mapping)
-
-            frame_worksheet.write_row(0, 0, ['ONNXLayer', 'TIDLDataID'])
-            frame_worksheet.write_row(0, 2, layers_diff[0].keys())
-            tidl_onnx_trace_mapping_keys = list(tidl_onnx_trace_mapping.keys())
-            for layer_idx in range(len(layers_diff)):
-                tidl_data_id = tidl_onnx_trace_mapping_keys[layer_idx]
-                tidl_onnx_trace_mapping_entry = tidl_onnx_trace_mapping[tidl_data_id]
-                frame_worksheet.write(layer_idx + 1, 0, tidl_onnx_trace_mapping_entry[2])
-                frame_worksheet.write(layer_idx + 1, 1, tidl_onnx_trace_mapping_entry[3])
-                frame_worksheet.write_row(layer_idx + 1, 2, layers_diff[layer_idx].values())
-            #
+        self._analyze_model_outputs(notidl_outputs_dirs, tidl_outputs_dirs, analyze_xlsx)
+        if self.kwargs['common.analyze_level'] > 0:
+            self._analyze_model_layers(notidl_outputs_dirs, tidl_traces_dirs, analyze_xlsx)
         #
         analyze_xlsx.close()
         print(f"INFO: Data successfully written to {analyze_xlsx_path}")
 
-    def get_traces(self, onnx_trace_folder, tidl_trace_folder, layer_info_path):
+    def _analyze_model_outputs(self, notidl_outputs_dirs, tidl_outputs_dirs, analyze_xlsx):
+        num_traces = len(tidl_outputs_dirs)
+        frame_worksheet = analyze_xlsx.add_worksheet('diff_outputs')
+        for frame_idx in range(num_traces):
+            golden_output_dir = notidl_outputs_dirs[frame_idx]
+            tidl_output_dir = tidl_outputs_dirs[frame_idx]
+            golden_output_files = [os.path.join(golden_output_dir, f) for f in os.listdir(golden_output_dir)]
+            tidl_output_files = [os.path.join(tidl_output_dir, f) for f in os.listdir(tidl_output_dir)]
+            golden_output_files.sort()
+            tidl_output_files.sort()
+            if len(tidl_output_files) == 0 or len(golden_output_files) == 0:
+                print(f"ERROR: Number of traces in notidl and tidl outputs do not match")
+                raise ValueError(f"ERROR: Number of traces in notidl and tidl outputs do not match")
+            #
+            num_outputs = len(tidl_output_files)
+            column_start_idx = 0
+            for output_idx in range(num_outputs):
+                tidl_output_file = tidl_output_files[output_idx]
+                golden_output_file = os.path.join(golden_output_dir, os.path.basename(tidl_output_file))
+                goldenBuffer = np.fromfile(golden_output_file, dtype=np.float32).flatten()
+                tidlBuffer = np.fromfile(tidl_output_file, dtype=np.float32).flatten()
+                diff_dict = self._get_tensor_diff(goldenBuffer, tidlBuffer)
+
+                frame_worksheet.write_row(0, column_start_idx+1, ['ONNXLayer', 'TIDLONNXLayer'])
+                frame_worksheet.write_row(0, column_start_idx+3, diff_dict.keys())
+
+                frame_worksheet.write(frame_idx + 1, column_start_idx+0, str(frame_idx))
+                frame_worksheet.write(frame_idx + 1, column_start_idx+1, os.path.basename(golden_output_file))
+                frame_worksheet.write(frame_idx + 1, column_start_idx+2, os.path.basename(tidl_output_file))
+                frame_worksheet.write_row(frame_idx + 1, column_start_idx+3, diff_dict.values())
+                column_start_idx += (3 + len(diff_dict))
+            #
+        #
+
+    def _analyze_model_layers(self, notidl_traces_dirs, tidl_traces_dirs, analyze_xlsx):
+        if len(notidl_traces_dirs) != len(tidl_traces_dirs):
+            print(f"ERROR: Number of traces in notidl and tidl folders do not match")
+            raise ValueError(f"ERROR: Number of traces in notidl and tidl folders do not match")
+        #
+        tidl_run_dir = os.path.join(self.run_dir, 'tidl')
+        num_traces = len(tidl_traces_dirs)
+
+        layer_info_dir = os.path.join(tidl_run_dir, 'artifacts', 'tempDir')
+        layer_info_files = [f for f in os.listdir(layer_info_dir) if f.endswith('layer_info.txt')]
+        layer_info_path = os.path.join(layer_info_dir, layer_info_files[0])
+        for frame_idx in range(num_traces):
+            frame_worksheet = analyze_xlsx.add_worksheet('diff_' + str(frame_idx))
+            tidl_onnx_trace_mapping = self._get_traces(notidl_traces_dirs[frame_idx], tidl_traces_dirs[frame_idx], layer_info_path)
+            layers_diff = self._get_layers_diff(notidl_traces_dirs[frame_idx], tidl_traces_dirs[frame_idx], layer_info_path, tidl_onnx_trace_mapping)
+
+            frame_worksheet.write_row(0, 1, ['ONNXLayer', 'TIDLDataID'])
+            frame_worksheet.write_row(0, 3, layers_diff[0].keys())
+            tidl_onnx_trace_mapping_keys = list(tidl_onnx_trace_mapping.keys())
+            for layer_idx in range(len(layers_diff)):
+                tidl_data_id = tidl_onnx_trace_mapping_keys[layer_idx]
+                tidl_onnx_trace_mapping_entry = tidl_onnx_trace_mapping[tidl_data_id]
+                frame_worksheet.write(layer_idx + 1, 0, str(layer_idx))
+                frame_worksheet.write(layer_idx + 1, 1, tidl_onnx_trace_mapping_entry[2])
+                frame_worksheet.write(layer_idx + 1, 2, tidl_onnx_trace_mapping_entry[3])
+                frame_worksheet.write_row(layer_idx + 1, 3, layers_diff[layer_idx].values())
+            #
+        #
+
+    def _get_traces(self, onnx_trace_folder, tidl_trace_folder, layer_info_path):
         """
         Computes the mapping between TIDL and ONNX traces
         """
@@ -233,18 +305,32 @@ class InferAnalyzeFinal(compile_base.CompileModelBase):
         #
         return tidl_onnx_trace_mapping
 
-    def get_layers_diff(self, onnx_trace_folder, tidl_trace_folder, layer_info_path, tidl_onnx_trace_mapping):
+    def _get_tensor_diff(self, goldenBuffer, tidlBuffer):
+        try:
+            delta = goldenBuffer - tidlBuffer
+        except:
+            delta = np.zeros_like(tidlBuffer)
+        #
+        eps = 1e-6
+        scale = np.mean(np.absolute(goldenBuffer))
+        scale = np.abs(scale)
+        abs_delta = np.absolute(delta)
+        max = np.max(abs_delta)
+        mean = np.mean(abs_delta)
+        mae = mean / scale if scale > eps else (mean+eps) / (scale+eps)
+        max_abs = np.mean(np.absolute(delta))
+        diff_dict = {'MeanAbsRelE': float(mae), 'MaxAbsE': float(max), 'MeanMaxAbsE': float(max_abs)}
+        return diff_dict
+
+    def _get_layers_diff(self, onnx_trace_folder, tidl_trace_folder, layer_info_path, tidl_onnx_trace_mapping):
         """
         Generate error summary for the entire network
         """
-        dropdown_list = list(tidl_onnx_trace_mapping.keys())
-        numLayers = len(dropdown_list)
-        mae_dict = {}
-        max_dict = {}
-        mae_abs_dict = {}
+        traces_files = list(tidl_onnx_trace_mapping.keys())
+        num_traces_files = len(traces_files)
         combined_diff = []
-        for idx in range(numLayers):
-            layer_idx = dropdown_list[idx]
+        for idx in range(num_traces_files):
+            layer_idx = traces_files[idx]
             tidl = os.path.join(
                 tidl_trace_folder,
                 tidl_onnx_trace_mapping[layer_idx][1],
@@ -257,26 +343,8 @@ class InferAnalyzeFinal(compile_base.CompileModelBase):
             tidlBuffer = np.fromfile(tidl, dtype=np.float32).flatten()
 
             tidl_onnx_trace_mapping[layer_idx].extend([goldenBuffer, tidlBuffer])
-
-            delta = []
-            try:
-                delta = goldenBuffer - tidlBuffer
-            except:
-                delta = np.zeros_like(tidlBuffer)
-            #
-            eps = 1e-6
-            scale = np.mean(np.absolute(goldenBuffer))
-            scale = np.abs(scale)
-            abs_delta = np.absolute(delta)
-            max = np.max(abs_delta)
-            mean = np.mean(abs_delta)
-            mae = mean / scale if scale > eps else (mean+eps) / (scale+eps)
-            max_abs = np.mean(np.absolute(delta))
-
-            mae_dict[layer_idx] = mae
-            max_dict[layer_idx] = max
-            mae_abs_dict[layer_idx] = max_abs
-            combined_diff.append({'MeanAbsRelE': float(mae), 'MaxAbsE': float(max), 'MeanMaxAbsE': float(max_abs)})
+            diff_dict = self._get_tensor_diff(goldenBuffer, tidlBuffer)
+            combined_diff.append(diff_dict)
         #
         return combined_diff
